### 2020-06-08 #2311 @whedon
[PRE REVIEW]: pyBEMT: An implementation of the Blade Element Momentum Theory in Python
**Submitting author:** @kegiljarhus (<a href="http://orcid.org/0000-0002-4144-0454">Knut Erik Teigen Giljarhus</a>)
**Repository:** <a href="https://github.com/kegiljarhus/pyBEMT" target ="_blank">https://github.com/kegiljarhus/pyBEMT</a>
**Version:** v0.1.0
