+++ joss/paper.bib
@@ -22,6 +22,7 @@
   title={AeroDyn theory manual},
   author={Moriarty, P.J. and Hansen, A.C.},
   year={2005},
+  doi = {10.2172/15014831},
   institution={National Renewable Energy Lab., Golden, CO (US)}
 }
 
@@ -49,6 +50,7 @@
   title={Free wake analysis of a wind turbine aerodynamics. Wind energy conversion. ASRL-TR-184-14},
   author={Gohard, J.C.},
   year={1978},
+  doi = {10.2172/5428795},
   publisher={Massachusetts Institute of Technology, Department of Aeronautics and Astronautics, Aeroelastic and
 Structures Research Laboratory: Cambridge, MA}
 }
