### 2020-06-17 #2311 @Kevin-Mattheus-Moerman
@kegiljarhus I have some brief comments on your paper. While we are looking for an editor to handle this submission I recommend that you look at the following.
1. Work on fixing references which do no render properly in the paper. Also add DOI's for all references that have them.
2. Consider citing papers that actually use this theory in design, to clarify importance of this work to research.
3. Check for typos (e.g. moomentum)
4. Consider, if possible/applicable, adding one or more graphical summary figures e.g. graphs or blade designs, to illustrate the functionality/application of this work.
5. This work may be flagged as ["Minor ‘utility’ package"](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements). It would help to clearly outline the novel/signigicant functionality offered by this work. How is it different/better than other packages, and how relevant is it to scientific research? Clearly sketch a picture of why would use and likely cite (because of the scientific significance) this work for their research.
