### 2020-06-17 #2311 @kegiljarhus
Thank you for the comments, @Kevin-Mattheus-Moerman, I will make some changes to the paper to better highlight the relevance of the package. Regarding the graphical summary, do you think adding the [two validation examples](https://pybemt.readthedocs.io/en/latest/validation.html) from the documentation would be appropriate, or were you more thinking about design examples?
