### 2020-06-17 72c7183
Add more references and examples to paper, along with fixing typos and references.
