+++ joss/paper.bib
@@ -55,3 +55,46 @@
 Structures Research Laboratory: Cambridge, MA}
 }
 
+
+@article{kotwicz2019rotor,
+  title={Rotor blade optimization and flight testing of a small UAV rotorcraft},
+  author={Herniczek, M.K. and Jee, D. and Sanders, B. and Feszty, D.},
+  journal={Journal of Unmanned Vehicle Systems},
+  volume={7},
+  number={4},
+  pages={325--344},
+  year={2019},
+  doi={10.1139/juvs-2017-0005},
+  publisher={NRC Research Press}
+}
+
+@article{borg2020,
+title = "A numerical performance analysis of a ducted, high-solidity tidal turbine",
+journal = "Renewable Energy",
+year = "2020",
+issn = "0960-1481",
+doi = "10.1016/j.renene.2020.04.005",
+author = "Borg, M.G. and  Xiao,Q. and Allsop, S. and Incecik, A. and Peyrard, C."
+}
+
+@article{bahaj2007power,
+  title={Power and thrust measurements of marine current turbines under various hydrodynamic flow conditions in a cavitation tunnel and a towing tank},
+  author={Bahaj, A.S. and Molland, A.F. and Chaplin, J.R. and Batten, W.M.J.},
+  journal={Renewable Energy},
+  volume={32},
+  number={3},
+  pages={407--426},
+  year={2007},
+  publisher={Elsevier},
+  doi={10.1016/j.renene.2006.01.012}
+}
+
+@article{theodorsen1937characteristics,
+  title={Characteristics of six propellers including the high-speed range},
+  author={Theodorsen, T. and Stickle, G.W. and Brevoort, M.J.},
+  journal={Annual Report-National Advisory Committee for Aeronautics},
+  pages={401},
+  year={1937},
+  publisher={US Government Printing Office}
+}
+
