+++ joss/paper.md
@@ -31,9 +31,9 @@ with a long history dating back to @glauert1935airplane.
 Despite the development of more sophisticated methods, such as vortex
 methods[@gohard1978free], the blade element momentum theory is still considered
 relevant for the study of rotor design. Its simple formulation lends itself to
-use in education and to quickly analyse new ideas. As examples of recent use
-of BEMT, it was used in [@kotwicz2019rotor] to compare airfoils and blade
-shapes for a UAV rotorcraft and in [@borg2020] together with computational
+use in education and to quickly analyse new ideas. As examples of recent uses
+of BEMT, it was used by @kotwicz2019rotor to compare airfoils and blade
+shapes for a UAV rotorcraft and by @borg2020 together with computational
 fluid dynamics to analyze the performance of a high-solidity tidal turbine
 design. Several open-source packages include an implementation of BEMT, for
 instance the `AeroDyn`[@moriarty2005aerodyn] solver used in the whole-turbine
