### 2020-07-02 #2311 @kegiljarhus
Thank you for the update @sjpfenninger . After the feedback from @Kevin-Mattheus-Moerman, I have updated the paper with some more references to use of BEMT, and also provided some examples of the use of the package. I hope this is satisfactory.
