### 2020-07-02 #2311 @sjpfenninger
Thanks @kegiljarhus ! What I'm still missing is a clearer differentiation between pyBEMT and other packages. I think this may just need some rewording. The paper implies that pyBEMT is the only software with a clean standalone implementation of BEMT, and possibly also the only Python package? I.e. -- why would somebody use pyBEMT rather than any of the numerous other packages? If you have more specific references to the research currently using pyBEMT, that would also be good to include to highlight the need for pyBEMT specifically as opposed to BEMT generally. Meanwhile bear with me as I look for reviewers.
