### 2020-07-02 #2311 @kegiljarhus
Thank you for the elaboration, I see what you mean. The key aspect is the combination of both propellers and turbines in one package, which is nice especially for educational purposes. With respect to propellers, it is also the only package I am aware of with a model for coaxial rotors. I will reword the paper to emphasize these points.

I am currently writing other papers where the package is used, so it's a bit of a chicken and egg situation as I would like to be able to cite the software used in those papers.
