### 2020-07-03 98cf688
Rewording of paper.
