### 2020-07-03 040ad74
Replace UAV rotorcraft by unmanned rotorcraft.
