### 2020-07-06 #2311 @sjpfenninger
> Thank you for the elaboration, I see what you mean. The key aspect is the combination of both propellers and turbines in one package, which is nice especially for educational purposes. With respect to propellers, it is also the only package I am aware of with a model for coaxial rotors. I will reword the paper to emphasize these points.

Great, if you can do that, that's sufficient then!
