### 2020-07-13 #2480 @sjpfenninger
@kegiljarhus just to remind you that some minor rewording of the paper is still waiting to happen as per your response in the pre-review issue:

> Thank you for the elaboration, I see what you mean. The key aspect is the combination of both propellers and turbines in one package, which is nice especially for educational purposes. With respect to propellers, it is also the only package I am aware of with a model for coaxial rotors. I will reword the paper to emphasize these points.
