### 2020-07-27 #2480 @kegiljarhus
> @kegiljarhus just to remind you that some minor rewording of the paper is still waiting to happen as per your response in the pre-review issue:

Dear @sjpfenninger . I apologize for the delay, I have been away on summer holiday. I did in fact make two commits with slight rewording just after your comment in the pre-review, they are included in the current article proof here. Please let me know if I should further clarify the statement of need before proceeding with the review.
