### 2020-08-12 #2480 @antviro
I have already my review report:
- I think it is a good idea to publish a BEM code in Python, despite there are already some free software which include such method (like the ones cited in the paper). A python BEM package allows its use in teaching and thesis which is interesting, moreover, this package also allows propellers to be simulated.
- Regarding its use in Wind Power, I think it would be interesting to add in the INI files some pitch variable to the case section ([case]) which would be summed afterwards to every pitch value (usually called twist) under the section [turbine]. So it would make easier its use to assess typical WT control.
- As API documentation states, airfoils are read by using AeroDYN/OpenFast type of file. Perhaps, it is interesting to mention in the paper (like airfoil file from XX is used as input) Anyway, I think the documentation could improve by including something about selecting another airfoil it in the "usage documentation", as well as by including some example of how to use another airfoil. Regarding the later, it would be nice that a local airfoil file can be used (like with ini files), for at the moment it seems users should include the airfoil file in a particular directory (in my case /usr/local/lib/python3.7/dist-packages/pybemt-0.1.0-py3.7.egg/pybemt/airfoils/ ) to use it.
