### 2020-08-12 kegiljarhus/pyBEMT#1 @rafmudaf
Installation docs
It would be more straightforward to show the installation procedures in two sequential steps:

1. Install the dependencies (i.e. the `pip` command you currently have included or by any other means)
2. Add pyBEMT to the Python path (i.e. `python setup.py install` or `pip install -e .`)

Its a somewhat minor detail but will make the installation process more complete. As is, it may not be clear to a novice user whether the `python setup.py install` command will install the dependencies.

This is in reference to the JOSS review: https://github.com/openjournals/joss-reviews/issues/2480.
