### 2020-08-13 c905ee4
Add optional twist angle to case. Updated tests and docs accordingly.
