+++ examples/tidal.ini
@@ -1,6 +1,7 @@
 [case]
 rpm = 220.0
 v_inf = 1.73
+twist = 0.0
 
 [turbine]
 nblades = 3
