### 2020-08-13 b330cbc
Rename solver test object to TestSolver.
