### 2020-08-13 75106b3
Update docs with information on how to add new airfoils.
