### 2020-08-13 97a4b16
Restructure the installation instructions.
