+++ README.rst
@@ -17,15 +17,7 @@ Notable features:
 Installation
 ------------
 
-The package can be installed using pythontools:
-
-.. code-block:: console
-
-        python setup.py install
-
-Or by ensuring that the pybemt directory is in the PYTHONPATH.
-
-The package requires the following libraries:
+To install the package, first ensure that the following required libraries are installed:
 
 - numpy
 - scipy
@@ -38,6 +30,14 @@ These can be installed using the provided ``requirements.txt``,
 .. code-block:: console
 
         pip install -r requirements.txt
+
+Next, the package can be installed using pythontools:
+
+.. code-block:: console
+
+        python setup.py install
+
+Alternatively, just add the pybemt directory to the PYTHONPATH.
 
 Documentation
 -------------
