### 2020-08-13 kegiljarhus/pyBEMT#1 @kegiljarhus
Thank you for the suggestion. This has been implemented in commit 97a4b1698462537218842415df5d3590e0b7ed90
