### 2020-08-13 #2480 @kegiljarhus
Thank you for the nice comments and constructive suggestions, @antviro. I have the following replies:

- Adding a global adjustment to the pitch is an excellent suggestion. This has been implemented in commit kegiljarhus/pyBEMT@c905ee45db6fc89dcb8a4f0ea6b9e41b6a2521e6. It may be confusing to mix pitch and twist in these definitions, so other suggestions for names are appreciated.
- I have added a description on how to add airfoils in commit kegiljarhus/pyBEMT@75106b395fd16e2b369a6deddc1e191f5b4c32bd . I want to keep the configuration and installation simple so I decided not to have a separate path for airfoils.
