### 2020-08-17 #2480 @sjpfenninger
Thanks @antviro for completing your review, can you confirm here if you are happy with the changes @kegiljarhus made in response to your comments?
