### 2020-08-20 #2480 @rafmudaf
My comments to the repository have been addressed, and I've finished reviewing. This tool is a great addition to the existing suite of simulation software for turbine and propeller aerodynamics. I recommend to publish.
