### 2020-08-20 #2480 @antviro
> * Adding a global adjustment to the pitch is an excellent suggestion. This has been implemented in commit [kegiljarhus/pyBEMT@c905ee4](https://github.com/kegiljarhus/pyBEMT/commit/c905ee45db6fc89dcb8a4f0ea6b9e41b6a2521e6). It may be confusing to mix pitch and twist in these definitions, so other suggestions for names are appreciated.

It is hard to suggest as the names are going to be used in different machines (propeller and turbines), I would write in the opposite way pitch for the global one, and twist for the one linked to each blade element (as it is usually called in wind power). However, I also consider it good as it is right now, as the explanation is clear enough, so that the user knows what is setting when changing any of the variables.
