### 2020-09-04 ff884c3
Minor fixes to paper.md
