### 2020-09-04 #2480 @sjpfenninger
@kegiljarhus Almost ready for acceptance! Can you check [my PR](https://github.com/kegiljarhus/pyBEMT/pull/2) for some minor fixing up of the paper. I notice that you do not provide a PyPI package. It might make sense to do so to make installation easier, but that's not mandatory.

After addressing the above, to proceed to acceptance, can you make a tagged release on GitHub and archive that on a repository such as  Zenodo or figshare, then report the version number and archive DOI here?
