### 2020-09-06 #2480 @kegiljarhus
Thank you, @sjpfenninger! I have merged your PR and tagged a v1.0 release.

Here is the Zenodo DOI: [10.5281/zenodo.4016951](https://zenodo.org/record/4016951)
