### 2020-09-07 #2480 @danielskatz
👋 @kegiljarhus - this looks almost ready to accept, which I am working on.  I see some reference issues that need to be fixed first, mostly by protecting cases with {}s:
 - Glauert has "Aerodynamic theory" which should be "Aerodynamic Theory" I think
 - in Gohard, should the "tr" in the middle of the tech report number be "TR"?
 - Hericzek et al. has "uav" in the title which I think should be UAV
 - Marten et al. - can you spell out IJETAE?
 - Silvestre et al. - please fix the cases on the conference title to be Title Case
