### 2020-09-07 09b0f11
Fix references in JOSS paper.
