+++ joss/paper.bib
@@ -2,7 +2,7 @@
 @article{marten2013qblade,
   title={{QBLADE}: an open source tool for design and simulation of horizontal and vertical axis wind turbines},
   author={Marten, D. and Wendler, J. and Pechlivanoglou, G. and Nayeri, C.N. and Paschereit, C.O.},
-  journal={IJETAE},
+  journal={International Journal of Emerging Technology and Advanced Engineering},
   volume={3},
   number={3},
   pages={264--269},
@@ -12,7 +12,7 @@
 @inproceedings{silvestre2013jblade,
   title={JBLADE: a propeller design and analysis code},
   author={Silvestre, M.A. and Morgado, J.P. and Pascoa, J.},
-  booktitle={2013 International Powered Lift Conference},
+  booktitle={2013 {I}nternational {P}owered {L}ift {C}onference},
   pages={4220},
   year={2013},
   doi={10.2514/6.2013-4220}
@@ -39,7 +39,7 @@
 @incollection{glauert1935airplane,
   title={Airplane propellers},
   author={Glauert, H.},
-  booktitle={Aerodynamic theory},
+  booktitle={Aerodynamic {T}heory},
   pages={169--360},
   year={1935},
   publisher={Springer Berlin Heidelberg},
@@ -47,7 +47,7 @@
 }
 
 @book{gohard1978free,
-  title={Free wake analysis of a wind turbine aerodynamics. Wind energy conversion. ASRL-TR-184-14},
+  title={Free wake analysis of a wind turbine aerodynamics. Wind energy conversion. {ASRL} {TR}-184-14},
   author={Gohard, J.C.},
   year={1978},
   doi = {10.2172/5428795},
@@ -57,7 +57,7 @@ Structures Research Laboratory: Cambridge, MA}
 
 
 @article{kotwicz2019rotor,
-  title={Rotor blade optimization and flight testing of a small UAV rotorcraft},
+  title={Rotor blade optimization and flight testing of a small {UAV} rotorcraft},
   author={Herniczek, M.K. and Jee, D. and Sanders, B. and Feszty, D.},
   journal={Journal of Unmanned Vehicle Systems},
   volume={7},
