### 2020-09-07 65e17f8
Another fix of references in JOSS paper.
