+++ joss/paper.bib
@@ -47,7 +47,7 @@
 }
 
 @book{gohard1978free,
-  title={Free wake analysis of a wind turbine aerodynamics. Wind energy conversion. {ASRL} {TR}-184-14},
+  title={Free wake analysis of a wind turbine aerodynamics. Wind energy conversion. {ASRL}-{TR}-184-14},
   author={Gohard, J.C.},
   year={1978},
   doi = {10.2172/5428795},
