### 2020-09-07 #2480 @kegiljarhus
:wave: @danielskatz, thank you for the suggestions, they are implemented now in the latest commit.
